Create tb_hazard_detection_unit.sv
probably one of the more difficult individual units to test imo. This testbench essentially creates a simplified processor that is only capable of addition but is pipelined and creates hazards for the hazard detection unit to hopefully detect. I have not run the testbench yet, as I am currently on my home computer and would rather do it at engineering hall to reduce lag, but I have commented it and it compiles.